+++ src/main/java/io/github/torand/fastersql/dialect/MariaDbDialect.java
@@ -81,6 +81,11 @@ public class MariaDbDialect implements Dialect {
     }
 
     @Override
+    public String formatCeilFunction(String operand) {
+        return "ceil(" + operand + ")";
+    }
+
+    @Override
     public boolean supports(Capability capability) {
         return SUPPORTED_CAPS.contains(capability);
     }