+++ src/main/java/io/github/torand/fastersql/dialect/OracleDialect.java
@@ -105,6 +105,11 @@ public class OracleDialect implements Dialect {
     }
 
     @Override
+    public String formatCeilFunction(String operand) {
+        return "ceil(" + operand + ")";
+    }
+
+    @Override
     public boolean supports(Capability capability) {
         return supportedCaps.contains(capability);
     }