+++ src/main/java/io/github/torand/fastersql/dialect/PostgreSqlDialect.java
@@ -92,6 +92,11 @@ public class PostgreSqlDialect implements Dialect {
     }
 
     @Override
+    public String formatCeilFunction(String operand) {
+        return "ceil(" + operand + ")";
+    }
+
+    @Override
     public boolean supports(Capability capability) {
         return SUPPORTED_CAPS.contains(capability);
     }